requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
pydantic==2.8.2
email_validator==2.2.0
python-dotenv==1.0.1
anyio==4.4.0
sniffio==1.3.1

